requirements.txt
Flask>=3.0.0
snowflake-connector-python>=3.6.0
python-dotenv>=1.0.0
waitress>=3.0.0
gunicorn>=21.0.0
Werkzeug>=3.0.0
